Extract caching from participant-state as a library
This will be used to keep a cache of values to cut on LF translation cost when serving transactions.

changelog_begin
changelog_end
