+++ ledger/participant-state/kvutils/BUILD.bazel
@@ -41,8 +41,8 @@ da_scala_library(
         "//ledger/metrics",
         "//ledger/participant-state",
         "//ledger/participant-state/protobuf:ledger_configuration_java_proto",
+        "//libs-scala/caching",
         "//libs-scala/contextualized-logging",
-        "@maven//:com_github_ben_manes_caffeine_caffeine",
         "@maven//:com_google_guava_guava",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:com_typesafe_akka_akka_actor_2_12",
