+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/caching/package.scala
@@ -3,8 +3,19 @@
 
 package com.daml.ledger.participant.state.kvutils
 
+import com.google.protobuf.MessageLite
+import com.daml.caching.{Cache, Weight}
+
 package object caching {
 
-  type Size = Long
+  implicit object `Bytes Weight` extends Weight[Bytes] {
+    override def weigh(value: Bytes): Cache.Size =
+      value.size().toLong
+  }
+
+  implicit object `Message Weight` extends Weight[MessageLite] {
+    override def weigh(value: MessageLite): Cache.Size =
+      value.getSerializedSize.toLong
+  }
 
 }
