+++ libs-scala/caching/BUILD.bazel
@@ -1,0 +1,20 @@
+# Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load(
+    "//bazel_tools:scala.bzl",
+    "da_scala_library",
+)
+
+da_scala_library(
+    name = "caching",
+    srcs = glob(["src/main/scala/**/*.scala"]),
+    tags = ["maven_coordinates=com.daml:caching:__VERSION__"],
+    visibility = [
+        "//:__subpackages__",
+    ],
+    deps = [
+        "@maven//:com_github_ben_manes_caffeine_caffeine",
+        "@maven//:org_scala_lang_modules_scala_java8_compat_2_12",
+    ],
+)
