+++ release/artifacts.yaml
@@ -128,6 +128,8 @@
   type: jar-scala
 - target: //libs-scala/auth-utils:auth-utils
   type: jar-scala
+- target: //libs-scala/caching:caching
+  type: jar-scala
 - target: //libs-scala/contextualized-logging:contextualized-logging
   type: jar-scala
 - target: //libs-scala/direct-execution-context:direct-execution-context
